extensions/wecom_bot/requirements.txt
dify_plugin==0.6.0b14
pycryptodomex==3.20.0
